Split zones.rs (768 lines) into subdirectory modules
Break the monolithic zones.rs into focused files under zones/:
- demand.rs (52 lines): ZoneDemand resource and methods
- stats.rs (109 lines): ZoneStats struct and gather_zone_stats
- market.rs (171 lines): vacancy helpers, market factors, demand computation
- systems.rs (73 lines): ECS systems (update_zone_demand, is_adjacent_to_road)
- tests.rs (368 lines): all unit tests
- mod.rs (28 lines): re-exports and ZonesPlugin

All pub items re-exported from mod.rs for backward compatibility.

Part of #1389
